+++ lib/models/api/v2/tweet_v2.dart
@@ -12,25 +12,56 @@ part 'tweet_v2.g.dart';
 /// Those need to be included in expansion parameters.
 @JsonSerializable()
 class TweetV2Response {
-  /// This field will contain the actual tweet, in this library we assume it will have exactly one element
-  final List<TweetV2> data;
+  /// This field will contain the actual tweet.
+  /// One endpoints of twitter API returns here a list an another returns a single object, hence Object
+  /// but in reality it should always be either TweetV2 or List<TweetV2>
+  /// If response is a list we always take the first element
+  final Object data;
 
   /// This field will contain any objects referenced from tweet inside data field, like media, users or tweets
   final TweetV2Includes includes;
 
-  const TweetV2Response({
+  /// This is the tweet extracted from [data] field
+  final TweetV2 tweet;
+
+  TweetV2Response({
     required this.data,
     this.includes = const TweetV2Includes(),
-  });
+  }) : tweet = _getTweet(data);
 
+  /// This function can accept both JSON that has a list and a single object in [data] field
+  /// So both
+  /// - {"data": ["created_at": "2020-09-18T18:36:15.000Z", "id": "1061967001177018368", ...
+  /// - {"data": {"created_at": "2020-09-18T18:36:15.000Z", "id": "1061967001177018368", ...
+  /// are valid
   factory TweetV2Response.fromRawJson(String str) =>
       TweetV2Response.fromJson(json.decode(str));
 
+  /// This function can accept both JSON that has a list and a single object in [data] field
+  /// So both
+  /// - {"data": ["created_at": "2020-09-18T18:36:15.000Z", "id": "1061967001177018368", ...
+  /// - {"data": {"created_at": "2020-09-18T18:36:15.000Z", "id": "1061967001177018368", ...
+  /// are valid
   factory TweetV2Response.fromJson(Map<String, dynamic> json) =>
       _$TweetV2ResponseFromJson(json);
 
   /// For the sake of this library we always assume there is exactly one tweet to be displayed
-  TweetV2 get tweet => data.first;
+  static TweetV2 _getTweet(Object data) {
+    /// If already proper object simply return it
+    if (data is TweetV2) {
+      return data;
+    }
+
+    /// Case for referenced tweets
+    if (data is List<TweetV2>) {
+      return data.first;
+    }
+
+    if (data is List) {
+      return TweetV2.fromJson(data.first as Map<String, dynamic>);
+    }
+    return TweetV2.fromJson(data as Map<String, dynamic>);
+  }
 }
 
 @JsonSerializable()